feat: Enhance student grade management system with additional functionality
- Added function to find and display the lowest score among student grades.
- Updated the user interface in `index.c` to include an option for finding the lowest score.
- Improved formatting and readability of output for displaying student grades.
- Refactored average value calculation to display with two decimal points.
- Enhanced clarity of output messages throughout the application.

+++ 01_programming_fundamentals/01_getting-started_programming_with_c/07_(development_projects)/02_student_grade_management_system/src/calculate_the_average_value.c
@@ -4,7 +4,6 @@
  *              and total values of student grades.
  */
 
-#include "../include/calculate_the_average_value.h"
 #include "../include/all_student_grades.h"
 #include <stdio.h>
 
@@ -26,6 +25,7 @@ int total_values() {
 }
 
 void calculate_the_average_value() {
-  mean = total_values() / all_score; // Calculate mean using total values
-  printf("Average Value: %d\n", mean); // Display the average value
+    mean = total_values() / all_score; // Calculate mean using total values
+    printf("Average Value: %.2f\n", (float)mean); // Display average with two decimal points
 }
+
